Move GameItem model class to the domain layer
