+++ lib/presentation/animal_game_screen/choice_a.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'package:flutter_kids_matching_game/models/gameItems.dart';
+import 'package:flutter_kids_matching_game/domain/game_item.dart';
 import 'package:flutter_kids_matching_game/presentation/animal_game_screen/animal_game_screen_controller.dart';
 
 // a left side choices with the name of objects like animals, colors ...
