+++ lib/presentation/animal_game_screen/choice_b.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:audioplayers/audioplayers.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import 'package:flutter_kids_matching_game/models/gameItems.dart';
+import 'package:flutter_kids_matching_game/domain/game_item.dart';
 import 'package:flutter_kids_matching_game/presentation/animal_game_screen/animal_game_screen_controller.dart';
 
 // a right side choices with the image of objects like animals, fruits ...
